requirements.txt
alabaster==0.7.12
amqp==2.4.2
argon2-cffi==18.1.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.4.0
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-appconf==1.0.3
django-braces==1.13.0
django-compressor==2.2
django-cors-headers==2.3.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-extensions==2.0.7
django-filter==2.1.0
django-map-widgets==0.2.2
django-model-utils==3.1.2
django-oauth-toolkit==1.2.0
django-redis==4.9.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
docutils==0.14
drf-rw-serializers==1.0.1
drf-yasg==1.10.0
factory-boy==2.11.1
Faker==1.0.4
flake8==3.5.0
future==0.17.1
geocoder==1.38.1
idna==2.8
imagesize==1.1.0
inflection==0.3.1
ipdb==0.11
ipython==7.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonfield==2.0.2
kombu==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
packaging==19.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
pluggy==0.6.0
prompt-toolkit==2.0.9
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.4.0
pytest==3.6.3
pytest-django==3.3.2
pytest-sugar==0.9.1
python-dateutil==2.8.0
python-slugify==1.2.5
python3-openid==3.1.0
pytz==2018.5
qrcode==6.0
ratelim==0.1.6
rcssmin==1.0.6
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
ruamel.yaml==0.15.91
six==1.12.0
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.1.0
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==3.3.1
